+++ amazon/iamidentitycenter/client/src/main/kotlin/org/http4k/connect/amazon/iamidentitycenter/util/OpenBrowser.kt
@@ -1,0 +1,3 @@
+package org.http4k.connect.amazon.iamidentitycenter.util
+
+typealias OpenBrowser = (String) -> Process